correlation count

+++ cpp/src/dmrpt/algo/drpt_global.hpp
@@ -16,7 +16,7 @@ namespace dmrpt {
 
     struct PriorityMap {
         int leaf_index;
-        int priority;
+        float priority;
     };
